pyproject.toml
[tool.poetry]
name = "transmep"
version = "0.1.0"
description = "Transfer learning for Mutation Effect Prediction"
authors = ["Tilman Hoffbauer <tilman.hoffbauer@rwth-aachen.de>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8,<3.11"
numpy = "^1.21"
tqdm = "^4.64"
torch = "1.12.1"
fair-esm = "^1.0.2"
arch = "^5.2.0"
captum = "^0.5"
plotly = "^5.7"
click = "^8.1"
fsspec = "^2022.3"
questionary = "^1.10"
aiohttp = "^3.8"
requests = "^2.27"
kaleido = "0.2.1"
scikit-learn = "^1.0"
deap = "^1.3.3"

[tool.poetry.dev-dependencies]
black = "^22.3"
isort = "^5.10"
pyflakes = "^2.4"
snakeviz = "^2.1.1"
pytest = "^7.1.2"
jupyterlab = "^3.4.4"
ipywidgets = "^7.7.1"
jupyterlab-widgets = "^1.1.1"

[tool.poetry.group.dev.dependencies]
liccheck = "^0.8.1"

[tool.liccheck]
authorized_licenses = [
    "bsd",
    "new bsd",
    "bsd license",
    "new bsd license",
    "simplified bsd",
    "apache",
    "apache 2.0",
    "apache software license",
    "apache software",
    "gnu lgpl",
    "lgpl with exceptions or zpl",
    "isc license",
    "isc license (iscl)",
    "mit",
    "mit license",
    "python software foundation license",
    "python software foundation",
    "zpl 2.1",
    "mozilla public license 2.0 (mpl 2.0)",
    "gnu library or lesser general public license (lgpl)",
    "historical permission notice and disclaimer (hpnd)",
]
unauthorized_licenses = [
    "gpl v3",
]
level = "CAUTIOUS"
dependencies = false
requirement_txt_file = "./requirements.txt"

[tool.liccheck.authorized_packages]
arch = "5.*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.poetry.scripts]
transmep = 'transmep.cli:main'
